# perf: auto-vectorize check_offsets `~-64-73%`